Create PATH
This simple script that I have not tested, is supposed to add a folder like bin in your [user's] home folder, and append's it to ~/.bashrc and also export's itself first. So short it does:

1. Export's PATH variable to /home/user/bin - For .sh .py .rb .java files you want to be running...
2. Because the above command is a session only meaning log out do it agian, this just add's it to the ~/.bashrc file so you only have to do one step instead of both of them.

This is probably nessicery on new fresh install's for which ever reason. If you would like me to make any more script's email me or message here or even on twitter.